requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
django-crispy-forms==1.6.0
django-pagedown==0.1.1
django-resized==0.3.5
Pillow==3.3.1
psycopg2==2.6.2
static3==0.7.0
django-autoslug==1.9.3




